+++ src/hooks/useMentor.js
@@ -1,7 +1,10 @@
 import { useMutation } from "@tanstack/react-query";
-import { createAssignment } from "../api/mentor";
-import { useContext } from "react";
-import { Context } from "../AppProvider";
+import {
+  createAssignment,
+  getAssignmentDetail,
+  getAssignmentList,
+  editAssignment,
+} from "../api/mentor";
 
 export const useCreateAssignment = () => {
   return useMutation({
@@ -17,6 +20,44 @@ export const useCreateAssignment = () => {
           error.response?.data?.detail || error.message
         }`
       );
+    },
+  });
+};
+
+export const useGetAssignmentList = () => {
+  return useMutation({
+    mutationFn: getAssignmentList,
+    onSuccess: (data) => {
+      console.log("과제 리스트 조회 성공", data);
+    },
+    onError: (error) => {
+      console.log("과제 리스트 조회 실패", error);
+    },
+  });
+};
+
+export const useGetAssignmentDetail = () => {
+  return useMutation({
+    mutationFn: getAssignmentDetail,
+    onSuccess: (data) => {
+      console.log("과제 상세 조회 성공", data);
+    },
+    onError: (error) => {
+      console.log("과제 상세 조회 실패", error);
+    },
+  });
+};
+
+export const useEditAssignment = () => {
+  return useMutation({
+    mutationFn: editAssignment,
+    onSuccess: (data) => {
+      console.log("과제 수정 성공", data);
+      alert("과제 수정이 완료되었습니다.");
+    },
+    onError: (error) => {
+      console.log("과제 수정 실패", error);
+      alert("과제 수정에 실패했습니다.");
     },
   });
 };
